feat(1794): S4 run_skill + skill_resolve op-backend delegation — closes #1794
Final stage of the skill-execution consolidation (#1794). Move the
run_skill + skill_resolve Control IR op LOGIC into reyn.skill while
keeping the op registration in op_runtime (P3/P4 registry↔backend
locality). Behavior-preserving rewire (not a package move): the op
handles stay registered + thin, delegating to the new backends.

- src/reyn/skill/run_skill.py (NEW): _resolve_skill_ref, _compute_skill_hash,
  + run(op, ctx, caller) — the moved handle body. ctx is duck-typed;
  OpContext/RunSkillIROp are TYPE_CHECKING-only → 0 executed reyn.runtime
  imports (layer-clean, the layer-direction gate auto-covers it).
- src/reyn/skill/skill_resolve.py (NEW): _categorize_source + resolve(...) —
  the moved handle body; same TYPE_CHECKING treatment.
- src/reyn/core/op_runtime/{run_skill,skill_resolve}.py: thinned to
  register(kind, handle) + a thin handle forwarding to the reyn.skill
  backend. The __init__ eager-import side-effect (registration) is unchanged.
- tests/test_op_backend_delegation.py (NEW): op-dispatch pin — registration
  stays in op_runtime (available_kinds) + live execute_op dispatch reaches
  the reyn.skill backend (skill_resolve resolves; run_skill resolution runs).
- Repoint the 3 unit tests importing the moved internals
  (_resolve_skill_ref / _compute_skill_hash) to reyn.skill.run_skill;
  update 2 eval-test docstring pointers.

Gate: layer gate PASS (reyn.skill executed reyn.runtime=0), replay green
(no MissingFixture = SP/tools unchanged), op-dispatch pins + repointed
units pass, ruff + tier-audit clean.

Refs #1794
