+++ tests/test_run_skill_form4_hallucination_recovery.py
@@ -31,7 +31,7 @@ from __future__ import annotations
 
 import pytest
 
-from reyn.core.op_runtime.run_skill import _resolve_skill_ref
+from reyn.skill.run_skill import _resolve_skill_ref
 from reyn.skill.skill_paths import SkillNotFoundError
 
 # ---------------------------------------------------------------------------
