+++ tests/test_skill_paths_eval_md.py
@@ -35,7 +35,7 @@ from reyn.stdlib.skills.skill_improver.copy_to_work_resolver_pure import (
 
 
 def _categorize_source(skill_dir: Path) -> str | None:
-    """Mirror of ``op_runtime.skill_resolve._categorize_source``.
+    """Mirror of ``reyn.skill.skill_resolve._categorize_source``.
 
     Replicated here so the test helper can synthesise the
     ``skill_resolve`` run_op output dict without standing up an
