+++ tests/test_skill_version_hash.py
@@ -22,10 +22,10 @@ from pathlib import Path
 import pytest
 
 from reyn.core.events.events import EventLog
-from reyn.core.op_runtime.run_skill import _compute_skill_hash
 from reyn.data.workspace.workspace import Workspace
 from reyn.schemas.models import RunSkillIROp
 from reyn.security.permissions.permissions import PermissionDecl, PermissionResolver
+from reyn.skill.run_skill import _compute_skill_hash
 from reyn.skill.sub_skill_runner import SubSkillResult
 
 # ---------------------------------------------------------------------------
